prettier everything

+++ contracts/StakeableUpgradeableV2.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.8.0;
 
-import "@openzeppelin/contracts-upgradeable/access/OwnableUpgradeable.sol";
+import '@openzeppelin/contracts-upgradeable/access/OwnableUpgradeable.sol';
 
 /**
  * @notice Stakeable needs to be inherited by other contracts for Staking capabilities
@@ -38,7 +38,7 @@ contract StakeableUpgradeableV2 is OwnableUpgradeable {
      *  stake amount for msg.sender
      */
     function _stake(uint256 amount) internal {
-        require(amount > 0, "Cannot stake nothing");
+        require(amount > 0, 'Cannot stake nothing');
         stakes[_msgSender()] += amount;
         emit Staked(_msgSender(), amount);
     }
@@ -48,7 +48,7 @@ contract StakeableUpgradeableV2 is OwnableUpgradeable {
      * unstake amount from msg.sender
      */
     function _unstake(uint256 amount) public {
-        require(stakes[_msgSender()] >= amount, "_unstake: exceeds balance");
+        require(stakes[_msgSender()] >= amount, '_unstake: exceeds balance');
         stakes[_msgSender()] -= amount;
         emit UnStaked(_msgSender(), amount);
     }
@@ -58,7 +58,7 @@ contract StakeableUpgradeableV2 is OwnableUpgradeable {
      * slash amount from msg.sender
      */
     function _slash(address account, uint256 amount) public onlyOwner {
-        require(stakes[account] >= amount, "_slash: exceeds balance");
+        require(stakes[account] >= amount, '_slash: exceeds balance');
         stakes[account] -= amount;
         emit Slashed(account, amount);
     }
@@ -78,38 +78,46 @@ contract StakeableUpgradeableV2 is OwnableUpgradeable {
      * bulk insert user's stake amounts.
      * totalAmount - By providing the totalAmount beforehand we can avoid crucial mistakes.
      */
-    function _bulkStake(User[] calldata users, uint256 totalAmount) internal onlyOwner {
-        require(users.length > 0, "bulkStake 0 users");
-        require(totalAmount > 0, "bulkStake 0 totalAmount");
-        uint calculatedAmount = 0;
+    function _bulkStake(User[] calldata users, uint256 totalAmount)
+        internal
+        onlyOwner
+    {
+        require(users.length > 0, 'bulkStake 0 users');
+        require(totalAmount > 0, 'bulkStake 0 totalAmount');
+        uint256 calculatedAmount = 0;
         for (uint256 i = 0; i < users.length; i++) {
-            uint amount = users[i].amount;
+            uint256 amount = users[i].amount;
             stakes[users[i].addr] += amount;
             calculatedAmount += amount;
             emit Staked(users[i].addr, amount);
         }
-        require(calculatedAmount == totalAmount, "incorrect totalAmount");
+        require(calculatedAmount == totalAmount, 'incorrect totalAmount');
     }
+
     /**
      * @notice
      * bulk slash user's stake amounts.
      * totalAmount - By providing the totalAmount beforehand we can avoid crucial mistakes.
      */
-    function _bulkSlash(User[] calldata users, uint256 totalAmount) internal onlyOwner returns(uint256) {
-        require(users.length > 0, "_bulkSlash 0 users");
-        require(totalAmount > 0, "_bulkSlash 0 totalAmount");
-        uint calculatedAmount = 0;
-        uint totalActuallySlashed = 0;
+    function _bulkSlash(User[] calldata users, uint256 totalAmount)
+        internal
+        onlyOwner
+        returns (uint256)
+    {
+        require(users.length > 0, '_bulkSlash 0 users');
+        require(totalAmount > 0, '_bulkSlash 0 totalAmount');
+        uint256 calculatedAmount = 0;
+        uint256 totalActuallySlashed = 0;
         for (uint256 i = 0; i < users.length; i++) {
-            uint amount = users[i].amount;
-            uint userStake = stakes[users[i].addr];
-            uint toSlash = amount > userStake ? userStake : amount;
+            uint256 amount = users[i].amount;
+            uint256 userStake = stakes[users[i].addr];
+            uint256 toSlash = amount > userStake ? userStake : amount;
             stakes[users[i].addr] -= toSlash;
             calculatedAmount += amount;
             totalActuallySlashed += toSlash;
             emit Slashed(users[i].addr, toSlash);
         }
-        require(calculatedAmount == totalAmount, "incorrect totalAmount");
+        require(calculatedAmount == totalAmount, 'incorrect totalAmount');
         return totalActuallySlashed;
     }
 }
